Fix : Capture 버그 수정
