+++ netlify.toml
@@ -1,0 +1,8 @@
+[build]
+  command = "npm run build"
+  publish = "dist"  
+
+[[redirects]]
+  from = "/*"
+  to = "/index.html"
+  status = 200
