+++ package.json
@@ -10,6 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "axios": "^1.7.9",
     "dom": "^0.0.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
