+++ src/App.jsx
@@ -5,8 +5,9 @@ import PostMain from './pages/PostRelayMainPage';
 import Footer from './components/Layout/footer';
 import SignupPage from './pages/SignupPage';
 import MyProfilePage from './pages/MyProfilePage';
-
 import CapturePage from './pages/CapturePage';
+import JoinRelay from './pages/JoinRelay';
+import RunningRelay from './pages/RunningRelay';
 function App() {
     return (
         <div className="App">
@@ -17,6 +18,8 @@ function App() {
                 <Route path="/signup" element={<SignupPage />} />
                 <Route path="/myprofile" element={<MyProfilePage />} />
                 <Route path="/capture" element={<CapturePage />} />
+                <Route path="/joinrelay" element={<JoinRelay />} />
+                <Route path="/runningrelay" element={<RunningRelay />} />
             </Routes>
             {/* <Footer /> */}
         </div>
